refactor: created getStrategicTypes module
